Updated UI
I implemented the backbone of the Expenses and Calculations scenes (All labels, text fields, buttons, etc.). I even included the table view on the expenses page already with a reference to the Expense class. I haven't defined what most of the buttons do yet, so all of the text fields and buttons (except for the Back and scene change buttons) are still useless for now. 

In short, all major UI elements are in place, however, they aren't performing functions yet.